add new default to overdimension heating in individual buildings
This allows them to cover heat demand peaks e.g. 10% higher than
those in the data.

The disadvantage of manipulating the costs is that the capacity is
then not quite right. This way at least the costs are right.

Doing it properly would require introducing artificial peaks, but this
creates new problems (e.g. what is going on with wind/solar/other
demand).